requirements.txt
aiohappyeyeballs==2.6.1
aiohttp==3.11.13
aiosignal==1.3.2
annotated-types==0.7.0
anyio==4.9.0
attrs==25.3.0
certifi==2025.1.31
charset-normalizer==3.4.1
frozenlist==1.5.0
h11==0.16.0
httpcore==1.0.9
httpx==0.28.1
idna==3.10
iniconfig==2.0.0
multidict==6.1.0
outcome==1.3.0.post0
packaging==24.2
pluggy==1.5.0
propcache==0.3.0
pydantic==2.10.6
pydantic_core==2.27.2
PySocks==1.7.1
pytest==8.3.5
python-dotenv==1.0.1
python-telegram-bot==22.0
requests==2.32.3
selenium==4.29.0
sniffio==1.3.1
sortedcontainers==2.4.0
trio==0.29.0
trio-websocket==0.12.2
typing_extensions==4.12.2
urllib3==2.3.0
webdriver-manager==4.0.2
websocket-client==1.8.0
wsproto==1.2.0
yarl==1.18.3
